Extract URL encoder components to new files
